pyproject.toml
[tool.poetry]
name = 'bx_django_utils'
version = "1alpha0"
description = 'Various Django utility functions'
authors = [
    'Jens Diemer <jens.diemer@boxine.de>',
    'Philipp Hagemeister <phihag@phihag.de>',
    'Florian Braun <mail@chiron-online.de>'
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Framework :: Django',
    'Intended Audience :: Developers',
]
license = 'MIT'
keywords=['django', 'utilities']
readme='README.md'

[tool.poetry.dependencies]
python = '>=3.6,<4.0.0'
bx_py_utils = '*'  # https://github.com/boxine/bx_py_utils
django = '*'
python-stdnum = "*"  # https://arthurdejong.org/python-stdnum/

[tool.poetry.dev-dependencies]
boto3 = "*"  # needed for e.g.: bx_django_utils/aws/ stuff!
requests-mock = "*"
tox = '*'
pytest = '*'
pytest-randomly = '*'
pytest-cov = '*'
pytest-django = '*'
flake8 = '*'
flynt = '*'
autopep8 = '*'
isort = '*'
poetry-publish = "*"  # https://github.com/jedie/poetry-publish
django-debug-toolbar = "*"

[tool.poetry.scripts]
publish = "bx_django_utils_tests.test_project.publish:publish"

[build-system]
requires = ['poetry>=0.12']
build-backend = 'poetry.masonry.api'


[tool.autopep8]
# https://github.com/hhatto/autopep8#pyprojecttoml
max_line_length = 120
exclude="*/htmlcov/*,*/migrations/*,*/volumes/*"


[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/config_files/#pyprojecttoml-preferred-format
atomic=true
line_length=120
case_sensitive=false
skip_glob=["*/htmlcov/*","*/migrations/*"]
multi_line_output=3
include_trailing_comma=true
known_first_party=["django_tools","django_tools_test_project"]
no_lines_before="LOCALFOLDER"
default_section="THIRDPARTY"
sections=["FUTURE","STDLIB","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
lines_after_imports=2


[tool.tox]
# https://tox.readthedocs.io/en/latest/example/basic.html#pyproject-toml-tox-legacy-ini
legacy_tox_ini = """
[tox]
isolated_build = True
envlist = py39-django{2.2,3.0,3.1},py38-django{2.2,3.0,3.1},py37-django{2.2,3.0,3.1},py36-django{2.2,3.0,3.1}
skip_missing_interpreters = True

[testenv]
passenv = *
whitelist_externals = make
deps =
    django2.2: django~=2.2.0
    django3.0: django~=3.0.0
    django3.1: django~=3.1.0
commands =
    django-admin --version
    make pytest-ci
"""
